Move SCM repository bootstrap into ScmManagerSetup
Move the post-SCM-Manager repository bootstrap logic into ScmManagerSetup,
because the bootstrap flow is only used as part of the internal SCM-Manager setup.

The bootstrap step now ensures the required remote repositories exist, initializes
the local workspace, aligns it with the remote main branch, recreates the local
directory structure, and pushes the generated bootstrap content.
